Refactor page path method to utilize custom page type paths
