+++ app/helpers/admin/url_helper.rb
@@ -1,5 +1,7 @@
-module Admin::UrlHelper
-  require 'uri'
+require 'uri'
+
+module Admin::UrlHelper  
+  include UrlHelper
 
   def format_path(path)
     return '' if path.to_s.empty?
@@ -21,7 +23,7 @@ module Admin::UrlHelper
   def build_url(base_url, page)
     return "#{base_url}#{page.path}" if default_route?(page)
 
-    path = lookup_page_path(page)
+    path = lookup_custom_page_path(page)
     return nil unless path
 
     "#{base_url}/#{path}/#{page.slug}"
@@ -34,16 +36,5 @@ module Admin::UrlHelper
     return "#{scheme}://#{host}:#{uri.port}" if host&.include?('localhost')
 
     "#{scheme}://#{host}"
-  end
-
-  def default_route?(page)
-    page_class_name = page.class.name
-    page_class_name == 'Page' || (defined?(DEFAULT_PAGE_TYPE_ROUTES) && DEFAULT_PAGE_TYPE_ROUTES.include?(page_class_name))
-  end
-
-  def lookup_page_path(page)
-    return nil unless defined?(CUSTOM_PAGE_TYPE_ROUTES) && CUSTOM_PAGE_TYPE_ROUTES.is_a?(Hash)
-
-    CUSTOM_PAGE_TYPE_ROUTES[page.class.name.to_sym]
   end
 end
