+++ lib/trusty_cms/version.rb
@@ -1,3 +1,3 @@
 module TrustyCms
-  VERSION = '7.0.28'.freeze
+  VERSION = '7.0.29'.freeze
 end
